+++ integration-test/exposure-test/src/main/java/com/truward/brikar/test/gossip/model/GossipChainMode.java
@@ -1,0 +1,12 @@
+package com.truward.brikar.test.gossip.model;
+
+/**
+ * @author Alexander Shabanov
+ */
+public enum GossipChainMode {
+  /** Tells this gossiper to call next element in the gossip chain. */
+  NEXT,
+
+  /** Tells this gossiper to be the end of the gossip chain. */
+  END
+}